requirements.txt
absl-py==0.10.0
aniso8601==8.1.1
astroid==2.4.2
astunparse==1.6.3
attrs==20.3.0
autopep8==1.5.4
bcrypt==3.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.21
dataclasses
decorator==4.4.2
dill==0.3.3
dm-tree==0.1.5
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.4.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphviz==0.15
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
isort==5.6.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kaggle==1.5.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lvis==0.5.3
lxml==4.6.2
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numpy
oauthlib==3.1.0
opencv-contrib-python
opencv-python
opencv-python-headless
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.0
promise==2.3
proto-plus==1.11.0
protobuf==3.13.0
psutil==5.7.3
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn
scipy
sentencepiece==0.1.94
six==1.15.0
SQLAlchemy==1.3.20
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow
tensorflow-addons
tensorflow-datasets
tensorflow-estimator
tensorflow-hub
tensorflow-metadata
tensorflow-model-optimization
termcolor
text-unidecode
tf-models-official
tf-slim
threadpoolctl
tifffile
toml
tqdm
typed-ast
typeguard
typing-extensions
uritemplate
urllib3
Werkzeug
wrapt
zipp
PyQt5
PyQt5-sip